Split Activation class into Activation_Flag and Activation_Notice, and ensure activation logic runs before init.
