+++ includes/Plugin.php
@@ -131,11 +131,17 @@ final class Plugin {
 		add_action( 'wp_head', $display_site_kit_meta );
 		add_action( 'login_head', $display_site_kit_meta );
 
+		$options = new Core\Storage\Options( $this->context );
+
+		// Register activation flag logic outside of 'init' since it hooks into
+		// plugin activation.
+		$activation_flag = new Core\Util\Activation_Flag( $this->context, $options );
+		$activation_flag->register();
+
 		// Initiate the plugin on 'init' for relying on current user being set.
 		add_action(
 			'init',
-			function() {
-				$options      = new Core\Storage\Options( $this->context );
+			function() use ( $options, $activation_flag ) {
 				$transients   = new Core\Storage\Transients( $this->context );
 				$user_options = new Core\Storage\User_Options( $this->context, get_current_user_id() );
 
@@ -151,6 +157,8 @@ final class Plugin {
 				$assets = new Core\Assets\Assets( $this->context );
 				$assets->register();
 
+				( new Core\Util\Reset( $this->context ) )->register();
+				( new Core\Util\Developer_Plugin_Installer( $this->context ) )->register();
 				( new Core\Util\Tracking( $this->context, $user_options ) )->register();
 				( new Core\REST_API\REST_Routes( $this->context, $authentication, $modules ) )->register();
 				( new Core\Admin_Bar\Admin_Bar( $this->context, $assets, $modules ) )->register();
@@ -160,7 +168,7 @@ final class Plugin {
 				( new Core\Notifications\Notifications( $this->context, $options, $authentication ) )->register();
 				( new Core\Util\Debug_Data( $this->context, $options, $user_options, $authentication, $modules ) )->register();
 				( new Core\Admin\Standalone( $this->context ) )->register();
-				( new Core\Util\Activation( $this->context, $options, $assets ) )->register();
+				( new Core\Util\Activation_Notice( $this->context, $activation_flag, $assets ) )->register();
 				( new Core\Util\Migration_1_3_0( $this->context, $options, $user_options ) )->register();
 				( new Core\Util\Migration_1_8_1( $this->context, $options, $user_options, $authentication ) )->register();
 
@@ -192,9 +200,6 @@ final class Plugin {
 				return $args;
 			}
 		);
-
-		( new Core\Util\Reset( $this->context ) )->register();
-		( new Core\Util\Developer_Plugin_Installer( $this->context ) )->register();
 	}
 
 	/**
